complete

+++ api/model/entity/page.go
@@ -1,6 +1,6 @@
 package entity
 
 type PageInfo struct {
-	TotalPages int32 `json:"total_pages"`
-	TotalCount int32 `json:"total_count"`
+	TotalPages int64 `json:"total_pages"`
+	TotalCount int64 `json:"total_count"`
 }
